Moved the auth function back into the React src, since they had nothing do with FB
