+++ functions/tsconfig.json
@@ -11,7 +11,5 @@
     "target": "es2017"
   },
   "compileOnSave": true,
-  "include": [
-    "src"
-  ]
+  "include": ["src", "../src/auth"]
 }
