+++ src/App.tsx
@@ -1,5 +1,3 @@
-import { useState } from "react";
-
 // import './App.css'
 import SignUpForm from "./components/SignUpForm/SignUpForm";
 import LoginForm from "./components/LoginForm/LoginForm";
